+++ task4/HomeTask4/HomeTask4.Web/Pages/Ingredients/Create.cshtml
@@ -17,7 +17,5 @@
     </form>
 </div>
 @section scripts{
-
-    <script src="~/lib/jquery-validation/dist/jquery.validate.min.js"></script>
-    <script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>
+    @{await Html.PartialAsync("_ValidationScriptsPartial"); }
 }
